Re-write BTreeIter to iterate rather than recurse.
The recursive version was a pain to reason about (and to write!). I
think it's clearer to make the stack of interior nodes explicit.

This has the advantage that we can now have a filter/comparison function
to short-circuit which nodes in the BTree we visit. Trying to add this
with the recursive version was painful, as each BTreeIter needed access
to the function.